# AV2/bazar-tem-tudo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs do BazarTemTudo",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movimentacoes": {
            "get": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Lista todas as movimentações",
                "operationId": "85344b365c8318d06a2b93c9acf9d97f",
                "responses": {
                    "200": {
                        "description": "Lista de movimentações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "quantidade": {
                                                "type": "integer"
                                            },
                                            "dataMovimentacao": {
                                                "type": "string"
                                            },
                                            "id_pedido": {
                                                "type": "integer"
                                            },
                                            "id_produto": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Registrar uma nova movimentação",
                "operationId": "7b3d3fdb8e23df0b3df9dc6a3d9f41d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "dataMovimentacao": {
                                        "type": "string"
                                    },
                                    "id_pedido": {
                                        "type": "integer"
                                    },
                                    "id_produto": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movimentação registrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "dataMovimentacao": {
                                            "type": "string",
                                            "example": "2024-06-02"
                                        },
                                        "id_pedido": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movimentacoes/{id}": {
            "get": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Buscar movimentação pelo ID",
                "operationId": "5f815bec3c88bc0c757c9007aae658a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da movimentação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimentação encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "dataMovimentacao": {
                                            "type": "string"
                                        },
                                        "id_pedido": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movimentação não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Atualizar uma movimentação pelo ID",
                "operationId": "b617ecf4490cfec0da453424bdc36569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da movimentação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "dataMovimentacao": {
                                        "type": "string"
                                    },
                                    "id_pedido": {
                                        "type": "integer"
                                    },
                                    "id_produto": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movimentação atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "dataMovimentacao": {
                                            "type": "string"
                                        },
                                        "id_pedido": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movimentação não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Remover uma movimentação pelo ID",
                "operationId": "5bfe4325e24cebed36130bfa89599aff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da movimentação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Movimentação removida com sucesso"
                    },
                    "404": {
                        "description": "Movimentação não encontrada"
                    }
                }
            }
        },
...
    },
...
}